ROFO-60 feat: 장소 검색 api 연결 + Location -> Place 이름 변경
