+++ data/src/main/java/com/weit2nd/data/model/search/TouristSpotsDTO.kt
@@ -1,0 +1,9 @@
+package com.weit2nd.data.model.search
+
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class TouristSpotsDTO(
+    @field:Json(name = "items") val items: List<TouristSpotDTO>,
+)
