--- data/src/main/java/com/weit2nd/data/model/tourism/TouristSpotsDTO.kt
@@ -1,9 +1,0 @@
-package com.weit2nd.data.model.tourism
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class TouristSpotsDTO(
-    @field:Json(name = "items") val items: List<TouristSpotDTO>,
-)
