+++ domain/src/main/java/com/weit2nd/domain/model/search/TourismType.kt
@@ -1,0 +1,6 @@
+package com.weit2nd.domain.model.search
+
+enum class TourismType {
+    TOUR,
+    FESTIVAL,
+}
