--- domain/src/main/java/com/weit2nd/domain/model/tourism/TourismType.kt
@@ -1,6 +1,0 @@
-package com.weit2nd.domain.model.tourism
-
-enum class TourismType {
-    TOUR,
-    FESTIVAL,
-}
