--- presentation/src/main/java/com/weit2nd/presentation/ui/select/location/SelectLocationIntent.kt
@@ -1,7 +1,0 @@
-package com.weit2nd.presentation.ui.select.location
-
-sealed class SelectLocationIntent {
-
-    data object SearchLocation : SelectLocationIntent()
-    data class StoreSearchWord(val input: String) : SelectLocationIntent()
-}
