--- presentation/src/main/java/com/weit2nd/presentation/ui/select/location/SelectLocationSideEffect.kt
@@ -1,4 +1,0 @@
-package com.weit2nd.presentation.ui.select.location
-
-sealed class SelectLocationSideEffect {
-}
