--- presentation/src/main/java/com/weit2nd/presentation/ui/select/location/SelectLocationState.kt
@@ -1,8 +1,0 @@
-package com.weit2nd.presentation.ui.select.location
-
-import com.weit2nd.domain.model.Location
-
-data class SelectLocationState(
-    val searchResults: List<Location> = emptyList(),
-    val userInput: String = "",
-)
