--- presentation/src/main/java/com/weit2nd/presentation/ui/select/location/map/SelectLocationMapSideEffect.kt
@@ -1,9 +1,0 @@
-package com.weit2nd.presentation.ui.select.location.map
-
-import com.kakao.vectormap.KakaoMap
-import com.kakao.vectormap.LatLng
-
-sealed class SelectLocationMapSideEffect {
-
-    data class MoveCamera(val map: KakaoMap, val position: LatLng) : SelectLocationMapSideEffect()
-}
