+++ presentation/src/main/java/com/weit2nd/presentation/ui/select/place/SelectPlaceIntent.kt
@@ -1,0 +1,7 @@
+package com.weit2nd.presentation.ui.select.place
+
+sealed class SelectPlaceIntent {
+
+    data object SearchPlace : SelectPlaceIntent()
+    data class StoreSearchWord(val input: String) : SelectPlaceIntent()
+}
