+++ presentation/src/main/java/com/weit2nd/presentation/ui/select/place/SelectPlaceSideEffect.kt
@@ -1,0 +1,4 @@
+package com.weit2nd.presentation.ui.select.place
+
+sealed class SelectPlaceSideEffect {
+}
