+++ presentation/src/main/java/com/weit2nd/presentation/ui/select/place/SelectPlaceState.kt
@@ -1,0 +1,8 @@
+package com.weit2nd.presentation.ui.select.place
+
+import com.weit2nd.domain.model.search.Place
+
+data class SelectPlaceState(
+    val searchResults: List<Place> = emptyList(),
+    val userInput: String = "",
+)
